Add single-project
Add single-project page with proper styles for mobile. Desktop yet to come because we are not too sure about the current outcome (lacks wow factor). Please note: Images are not yet committed. I don't know if they will in the future, but they will surely need to be better organized.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>