Add mammoccio & social
Add the mammoccio template, the relative (temporary) implementation file emanuele.kit, styles, and the relative .gitignore directive for the compiled html.

This commit also adds the social thing with styles and icon-fonts. It has been put into the Footer css group because it will be also put in there and there it belongs.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>