Abstract content from markup
IMPORTANT UPDATE: The markup now gets filled dynamically by CodeKit. All content for projects and mammocci is stored inside the /contents/ folder and organized specifically for the single templates. The templates then are imported into the name.kit files together with all other components (header, footer, ecc) and everything gets spit out onto /htdocs/.

This lets me reuse and combine bits of information in order to output a clean and semantic web page.

There is minor bug related to the fact that at the moment CodeKit does not consent to assign the content of a variable to another variable ($varB = $varA). The bug is now present only for projects meta tags, as far as I know, so it's nothing major.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>