Some footer magic
Now the footer uses flexbox in order to achieve the clean look the designer imagined. It has one block on the left and two on the right, distributed nicely according to the width of the body. The elements also get automatically distributed vertically so that they cover the whole height. Flexbox really is magical.

There is also the map link feature: Apple Maps app is opened on iOS and OSX otherwise the link goes towards Google Maps (this might need some testing, because I might think of switching to a geo: uri schema in order to open native apps on Android devices).

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>