New home and new centering behavior
New styles appearance for the home, particularly in the #about and #mammocci sections.

This commit also includes a new behavior for centering page blocks in the viewport. Now centered blocks has a .centered class, while before the whole <body> was centered. This is to allow the #about section to cover the whole screen width and of course it might be used for other elements as well in the future.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>