Smarter header shrinking
The header now starts shrinking as soon as the the user scrolls and the .shrinked class is removed when the top of the page is 30px away. This feels very fluid and cool. I also applied a translateZ(0) hack to the header to increase performance and the transparency is applied only when the header is shrinked.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>