Improves animation performance on homepage and adds real content
Scrolling on homepage was pretty jaggy on Chrome on retina displays due to the heavy animations on hover. I have improved this by adding a delay on transitions so that animations don't get triggered while casually scrolling, but only when the user is focusing on something. I also tried to tap into hardware acceleration with the use of some 3d transforms.

This commit also includes a few minor fixes, some content (removing lorem ipsum) on the homepage and some code polishing.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>