Smooth scrolling to contacts
Clicking on the Contacts link in the nav now starts a smooth scrolling rather than an abrupt movement. With the previous implementation the user easily lost context because all other links in the navigation bar bring to another page. This helps the user understand that he is still on the same page. Better context, better usability. Nice indeed.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>