Improves the scrollTo function for better cross-browser support
I was having troubles animating document.body (working in webkit) and document.documentElement (working in Firefox) in the scrollTo function, so I have abstracted the thing and checked to see if we are on FF or not.

I know this might not be ideal, but I couldn't figure out why this was happening, nor could I find any information online.

This will need testing on IE.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>