Add super smart font-family declaration
This will need a whole blog post to explain. It is a very smart way of using the font-family declaration. Basically we will always try and display the best sans-serif font available, mainly Helvetica Neue Light on Macs and Segoe UI on Windows.

The blog post is already well underway and will be published soon, presumably on Gravida's Google+ page.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>