Add fade-it-in support
The fade-it-in functionality is now in place. It will fade in selected elements when they get in the viewport. This will only works for viewportWidth > 1025.

This commit also disabled the shrinking header functionality on tablets, because scroll events were misfiring on touch-enabled devices.

The scirpts file has also been thoroughly reorganized to have better performance and functionality. Now the most relevant viewport-related scripts fire each time the window gets resized.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>