Add language translation feature
Provision translations. English is there and working. For Italian, we are missing a few pages and it's highly unlikely that I will work on them. This commit serves the purpose of having a clean working directory without losing any progress. Unfortunately, I didn't use feature branches at the time, so it would be very difficult to save the language feature on a separate branch. This will do for now.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>